Refactor: fencer, libstonithd: Bounds-check stonith-watchdog-timeout
There may be cases that aren't technically correct (especially in
conjunction with very large SBD_WATCHDOG_TIMEOUT values) -- but with
timeout values higher than INT_MAX, there's no practical relevance. A
timeout of INT_MAX milliseconds is over three weeks if I did the math
correctly.

The most important thing is to avoid overflow to negative values in
configurations with unreasonably large timeouts.

+++ daemons/fenced/pacemaker-fenced.c
@@ -42,7 +42,7 @@
 #define SUMMARY "daemon for executing fencing devices in a Pacemaker cluster"
 
 char *stonith_our_uname = NULL;
-long stonith_watchdog_timeout_ms = 0;
+long long stonith_watchdog_timeout_ms = 0;
 GList *stonith_watchdog_targets = NULL;
 
 static GMainLoop *mainloop = NULL;
